\name{append-methods}
\docType{methods}
\alias{append}
\alias{append-methods}
\alias{append,ANY,ANY-method}
\alias{append,StringSet,character-method}
\title{Add strings to a StringSet.}
\description{
 This method appends or adds one or more strings
 to an existing \code{\link{StringSet-class}}.

}
\section{Methods}{
\describe{
This method is for appending strings to a
\code{\link{StringSet-class}} object
after it has been constructed. 
\item{x = "StringSet", values = "character"}{
   the first argument is the \code{\link{StringSet-class}}
 object to which the values are to be added, and the second is the
 strings that are being added.
}
}}
\keyword{methods}

